// src/client/models/departure.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DisplayInformation } from './display-information';
// May contain unused imports in some cases
// @ts-ignore
import type { Link } from './link';
// May contain unused imports in some cases
// @ts-ignore
import type { Route } from './route';
// May contain unused imports in some cases
// @ts-ignore
import type { StopPoint } from './stop-point';

export interface Departure {
    'display_informations'?: DisplayInformation;
    'route'?: Route;
    'stop_point'?: StopPoint;
    'stop_date_time'?: string;
    'links'?: Array<Link>;
}